refactor - file_switcher now in own file
also removed all (?) references to filelist
