--- spyderlib/widgets/html_delegate.py
@@ -1,36 +1,0 @@
-from PyQt4 import QtCore, QtGui
-
-class HTMLDelegate(QtGui.QStyledItemDelegate):
-    """ Taken from http://stackoverflow.com/a/5443112/2399799
-    """
-    def paint(self, painter, option, index):
-        options = QtGui.QStyleOptionViewItemV4(option)
-        self.initStyleOption(options,index)
-
-        style = QtGui.QApplication.style() if options.widget is None else options.widget.style()
-
-        doc = QtGui.QTextDocument()
-        doc.setHtml(options.text)
-
-        options.text = ""
-        style.drawControl(QtGui.QStyle.CE_ItemViewItem, options, painter);
-
-        ctx = QtGui.QAbstractTextDocumentLayout.PaintContext()
-
-        textRect = style.subElementRect(QtGui.QStyle.SE_ItemViewItemText, options)
-        painter.save()
-        painter.translate(textRect.topLeft())
-        painter.setClipRect(textRect.translated(-textRect.topLeft()))
-        doc.documentLayout().draw(painter, ctx)
-
-        painter.restore()
-
-    def sizeHint(self, option, index):
-        options = QtGui.QStyleOptionViewItemV4(option)
-        self.initStyleOption(options,index)
-
-        doc = QtGui.QTextDocument()
-        doc.setHtml(options.text)
-        doc.setTextWidth(options.rect.width())
-        return QtCore.QSize(doc.idealWidth(), doc.size().height())
-
